+++ apps/expo-example/src/wagmi/config.native.ts
@@ -16,11 +16,8 @@ export const wagmiConfig = createConfig({
       projectId: ZERODEV_PROJECT_ID,
       chains,
       rpId: RP_ID,
-      fetchOptions: {
-        // RP_ID needs to be the same as Origin
-        headers: { Origin: `https://${RP_ID}` },
-      },
       apiKeyStamper: createSecureStoreStamper(),
+      // Optional: omit passkeyStamper if your app doesn't use passkey auth.
       passkeyStamper: createReactNativePasskeyStamper({ rpId: RP_ID }),
       sessionStorage: asyncStorageAdapter,
       persistStorage: asyncStorageAdapter,