+++ packages/react-kit/CHANGELOG.md
@@ -1,6 +1,22 @@
 # @zerodev/wallet-react-kit
 
+## 0.0.1-alpha.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @zerodev/wallet-core@0.0.1-alpha.22
+  - @zerodev/wallet-react@0.0.1-alpha.25
+
 ## 0.0.1-alpha.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @zerodev/wallet-core@0.0.1-alpha.21
+  - @zerodev/wallet-react@0.0.1-alpha.24
+
+## 0.0.1-alpha.0
 
 ### Patch Changes